Update patch set 1
Patch Set 1:

hi roland,

thanks for the nice welcome and sorry for the delayed response.

at sap we are currently doing a poc for using monasca in a production environment, if we get it it working and stabilized.

thanks for your comments in the code - i'll try to adjust the coding according to your suggestions - maybe tomorrow. all you suggestions make sense to me.

i tested this code in our environment with a non Default domain and as well in our docker based testing env without the additional options set and with them left empty and it all works well. is there an additional need to test it with the vagrant minmon env then as well too?

best wishes and looking forward to bring in more fixes in the future - t

Patch-set: 1
Label: Verified=0