Update patch set 2
Patch Set 2:

I don't know hot to make blueprint. I only observed that in your code was two implementations of connection to database (raw mysql and hibernate). Also I saw that fixes was added only to raw mysql part. I believe that hibernate could have some performance issue or something (I'm not sure what cause is connected that you have still mysql and hibernate). But I think that it is good to have support to many databases. And I believe that jOQQ is better for this goal than hibernate. In jOQQ you could create almost every query like as in raw connection, but in jOQQ is possible to write one query which should work on several databases. If you don't want this code it's ok for me, but in monasca-api I added also similar solution for python part.

Patch-set: 2
Label: Verified=0