Improving kafka performance
The kafka client library recieves batches of a certain size from kafka
regardless of the batch size I ask for.  Asking for a batch size that's larger
than the respose size results in multiple requests to kafka until the batch
size I asked for is recieved.

Asking for a single message still causes the kafka client to recieve a batch of
data from kafka but doing it this way, which is the same way that the kafka
library __iter__ function works, results in a ~5x improvement in throughput
over asking for larger batch sizes.

Change-Id: I3380df56749a577ae7116e5da841dcb91c85312a