Improving kafka performance
The kafka client library recieves batches of a certain size from kafka
regardless of the batch size I ask for.  Asking for a batch size that's larger
than the respose size results in multiple requests to kafka until the batch
size I asked for is recieved.

Asking for a single message still causes the kafka client to recieve a batch of
data from kafka but doing it this way, which is the same way that the kafka
library __iter__ function works, results in a ~5x improvement in throughput
over asking for larger batch sizes.

Increasing the requested batch size results in a dramatic increase in
performance.  For the Java consumer the default read size is 1MB compared to
the 4KB default read size in the Python consumer.  Increasing the Python
consumer to match the Java version results in a ~10x improvement.

Change-Id: I3380df56749a577ae7116e5da841dcb91c85312a