Sporadic and deterministic support
Commit provides sporadic support for metrics.
Metrics can be send right now with sporadic keyword.

Alarms can be created with deterministic keyword.
'deterministic' can be passed as part of alarm expression using
one of following forms:

 * keyword deterministic
 * deterministic={true,yes,1}
 * deterministic={false,no,0}

If 'deterministic' is not found in expression it is assumed
to be 'true'.

Implements: blueprint alarmsonlogs
Change-Id: Ia42f9a1be37c31416bdac341b092fe527f860c16