Update patch set 1
Patch Set 1:

Joe,

python projects in OpenStack use a common setup to build documentation. That setup uses RST files and sphinx - and the standard directory is doc/source for the content.

this repository has nothing like that, so the test requirements you use for sphinx are completely unused.

I only wanted to remove sphinxcontrib-docbookrestapi since that is a dead requirement - and then noticed the rest.

So, whenever you want to add documentation again, create RST files under doc/source and add as minimal setup the following to test-requirements:
oslosphinx
sphinx>=1.1.2,!=1.2.0,<1.3

The rest is not needed in general - you can add these if you have very specific requirements on a case by case base.

Patch-set: 1