Update patch set 2
Patch Set 2:

It was not the case that we have found an error with default partitioner, there were another reasons:
- murmur2 mimics the behaviour that is implemented in Java (+1)
- default one in current (0.9.5) version is deprecated (+1) because of https://github.com/dpkp/kafka-python/issues/374

If I was changing that, I though it would be nice to add round robin here as an option.

As for your comment, that makes me think actually, if that's really a good idea. I am not sure how long it will take before moving to 1.0.0, but once it it upgraded we might want to revisit KafkaProducer and modify it.

Patch-set: 2
Label: Verified=0