Update patch set 19
Patch Set 19:

I think all the code is fine, but wondering about switching the name from "sporadic" to "no_undetermined" or an even better name. Yeah, I know, I'm not a popular guy right now.

With the new change all that is being done is preventing the alarm transitioning to the UNDETERMINED state. The new parameter can be used for both periodic or sporadic metrics, so calling the parameter 'sporadic" is a little misleading.

Patch-set: 19
Label: Verified=0