Update patch set 20
Patch Set 20:

I think sporadic is an excellent name for describing metrics that are event based. 

But, I thought the way that this had evolved is that the sporadic parameter on the alarm sub-expression is an attribute of the alarm definition and not an attribute of the metric anymore. All that is being done with the sporadic parameter is that it is eliminating the alarm from transitioning to the UNDETERMINED state. Additionally, the parameter can be applied to both periodic or sporadic metrics. Since it is applicable to both periodic and sporadic metrics, that is why I think continuing to call is sporadic is a bit misleading. I think the intended use case is still for sporadic metrics, but it can be used for both periodic and sporadic.

I think you are viewing the sporadic parameter as supplying additional information to the alarm about the metric, as in telling the alarm sub-expression, that this metric is a sporadic metric and as a result, don't transition to UNDETERMINED. I'm just viewing it as eliminating the UNDETERMINED state, which means it is applicable to any metric. 

In either case, I think the functionality is the same as far as I can tell.

Patch-set: 20
Label: Verified=0