Update patch set 20
Patch Set 20:

Ok, so maybe for metric let's stay with sporadic
and for alarm I'd go with deterministic. It is rather long, I know.
But let's assume

 deterministic=true

means that in every moment of time it should be possible to establish the state alarm should be in. If it wouldn't be possible (i.e. missing metrics) alarm should transition to UNDETERMINED state. On the other hand

 deterministic=false

would mean exact opposite meaning that it is not always possible to establish that state, so we assume (according to implementation in threshold) that only possible state (apart from ALARM) is OK.

What about that idea ?

Patch-set: 20
Label: Verified=0