Update patch set 20
Patch Set 20:

I'd go with deterministic, since you like it (I hope, in time you'll love it). For me, in context of alarm's expression - single word is the best thing, because string being alarm's expression is kept as short as possible.

Tristate seems to be fine, but if there would be a new state in the future, it'd not make sense IMHO. The last one is too exotic, I guess - but made me laugh my lungs out.

Ok, so I am going with deterministic, because I'd really love to close this change down and proceed further. Apart from thresh I have also monasca-ui to modify and that much time. If someone would have any objections for deterministic, just -1 this review with comment. I guess that is fair enough for now.

Patch-set: 20