Update patch set 29
Patch Set 29:

I know it's late in the process, but I wonder if putting 'deterministic' as a keyword in front of the function would make it clearer. Putting it inside the function, between the metric name and period makes it seem like its a part of the aggregation function when it really applies to the entire sub expression.

ex: "deterministic max(cpu.idle_perc, 1) > 10 AND avg(cpu.user_perc) < 10"

Patch-set: 29