Update patch set 2
Patch Set 2:

Should this review be abandoned. 

I think adding a "sporadic" field to a metric could potentially be useful, if it was stored in the metrics DB (Vertica or InfluxDB). Then, when a client queries the metrics resource they could determine whether the metric was sporadic or not, which could be useful to the operator when creating alarm definitions that need to use the "deterministic" parameter. If the sporadic field was stored it could also be returned.

However, I don't think there is any additional or different processing that occurs if a metric has the sporadic field specified at this point. The processing in the threshold engine is purely based on whether the deterministic option is set to true/false in the alarm definition.

This review is modifying the metrics definition in the MySQL DB, but I'm not sure how that will be useful. It could have been potentially used by the threshold engine, but with the deterministic change it isn't used. It also isn't returned when querying the metrics resource in the API.

Patch-set: 2
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0