Update patch set 29
Patch Set 29:

(1 comment)

I see your point that in current shape it looks like it is used in aggregation when it is actually not. Just, if you would actually have deterministic keyword in front of the function that would make expression harder to read. Having it where it is right now - is more or less an attribute of expression and we can think of it that way, just as period is an attribute. 

There's also another point that I think is valid. You've mentioned that deterministic applies to entire sub expression. Yes it does. Just as period, periods, operator and threshold or even function. All of them makes expression fully defined and sometimes pair of them (like period and periods) only together mean something however they are not next to each other.

Patch-set: 29
Label: Verified=0