Update patch set 29
Patch Set 29:

I agree with checking if the deterministic key exists or not. It is
simper solution. But for putting 'deterministic' in front of the function, I don't think this is the best solution. For me more readable
and clearer way is to keep this keyword between metric name and period.

Patch-set: 29
Reviewer: Gerrit User 20033 <20033@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0