Update patch set 2
Patch Set 2:

I'm not a time series database nor a monasca experto, so, sorry if it's a silly question, but how can we store the 'sporadic' field to the metrics DB (Vertica or influxDB)? Supose we are trying to get this information to show in the dashboard, and the user sent the first measurement as the metric being sporadic, but didn't specified in the other measurements if it is sporadic, how may we retrieve this information? Will we have to look in all of the database entries for this metric?

Patch-set: 2
Reviewer: Gerrit User 15359 <15359@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0