Update patch set 33
Patch Set 33:

Hi Craig,

I've fixed that test, it got somehow lost I guess.

As for your proposition, I think I am ok with it. I had the same plan in my head. So if you don't have any objections we can start with submitting monasca-common and then monasca-thresh and so on.

As for Hibernate, well SessionFactory won't be able to build Session if model does not match database (at least in this case where it is told not to override database). So unless 'sub_alarm_definition' defines new column it won't work. However Hibernate is disabled in devstack environment so that should not be an actual issue and all tests will go through native MySQL implementation which does not require any Pojo classes.

As for the tempests test, I assume you would like to see some more apart from those that I've added to cover creating new type of alarms ?

Patch-set: 33