Update patch set 3
Patch Set 3:

The question I have is why is this being passed along as the partition key?  I would have thought that you would have a hash or an integer representation of your target partition by the time you sent the data to Kafka.  I'm not sure that allowing arbitrary information to be turned into the key on the Kafka side is a good idea.

Patch-set: 3