Update patch set 3
Patch Set 3:

The partition key can only be a sequence of bytes which is why the unicode type did not work.  Encoding that as utf-8 does work but looking at the log publisher code it looks like the key could be quite large.  

In the Monasca API we do something similar but we hash the tenant id and dimension set and use that as the key so that we have a known size for the key.

In the Monasca API there was originally a desire to have each measurement go to a consistent partition.  We no longer have that capability in the Python version of the Monasca API.  Is that something you need in the log API?

Patch-set: 3