Update patch set 3
Patch Set 3:

The reason why the Java monasca-api used a specific key was to guarantee that the same metrics went to the same consumer, for the purposes of the Monasca Anomaly engine. The old Monasca Anomaly engine was not a clustered engine and therefore you had to guarantee that the same metrics ended up going to the same consumer/process. We were only computing anomalies on a per metric basis.

Currently, the path forward for anomaly detection is using Monasca Analytics and that uses Spark and Spark is clustered. 

My understanding of LogStash, which is not the best, is that it is stateless. Each log message is processed independently of any other state. Our intended usage is to output a metric based on some per-log message criteria, not based on a stream or other log messages. If that is accurate, then I don't think we need a special key to provide specific guarantees on what partitions messages are stored with and what consumers process them. Probably, the best parition key is one that does a uniform distribution of log messages to all paritions.

I think we can use a timestamp as the key, or whatever the default method is.

Patch-set: 3
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0