Update patch set 3
Patch Set 3:

Craig, the tempest test fail because they refer to the current (old) monasca packages versions (monasca-common:1.1.0).

If my understanding is correct, we should first merge this change. Then monasca-thresh and monasca-persister, because they depend on monasca-common.
As the last one we should merge monasca-api for tempest tests to pass with the new package versions.

Is it correct?

Patch-set: 3