Fix crash-loop in thresholder caused by floating point numbers used as dimension filter (e.g. avg(some_timer{quantile=0.99}) > 0.1)
* add DECIMAL (floating point value) to allowed tokens for dimensions
* use antlr 4.2.2 instead of 4.2.0

Without this fix, an alarm-definition like the above one will break the Java-ORM layer because the alarm-definition (submitted using the Python API) cannot be parsed by the generated ANTLR parser. As a consequence the thresholder worker will terminate, leading to a crash-cycle. I could not find a way to make the ORM-layer ignore non-parseable alarm-definitions. So maybe this fix should be complemented by an extension of the Python API that uses exactly the same ANTLR grammar to validate incoming alarm-definitions.

Change-Id: Ibbd41e1b817ebadc74b1b4a677db898cfa173ccb