Update patch set 9
Patch Set 9:

Hi Joe, https://review.openstack.org/#/c/351429/ has merged and the review to add pykafka to global requirements has also merged.

I think this is ready to go too, but please let me know if there is anything else that we are waiting on. If this is ready, I'll merge it.

Patch-set: 9