Update patch set 3
Patch Set 3:

Tomasz, thanks for catching the style issue. Not sure what happened, I set the code template to 2 spaces. patch 3 fixed the space issue.

The configurations in this class are Cassandra only and not available in other places. I am following the precedence of influxdb to put this in the common. We could just leave it in the persister, six and half a dozen, does the same thing.

Patch-set: 3
Label: Verified=0