Update patch set 4
Patch Set 4:

I will put this comment here :D.

All of the changes that are about adding cassandra should have the same time. The changes in API + Persister (if there any) should have ```Depends-On``` on this change to ensure that monasca-common with the cassandra configuration is used in there.

I didn't check other changes at the moment for that, if it is already there sorry, if not - please add.

Patch-set: 4