Retry publish once on failures
Clear the topic metadata in case the IP Addresses have changed.
This can be common when Monasca is run in Kubernetes and the
Kafka pod is restarted.

This changes the actual Kafka library code that was saved into
this repo. It might be possible to use newer code, but that
would introduce lots of other changes so making the simplest
fix here.

Change-Id: If48971c40883b5be10adec947562cdda7e82d77c
Story: 2001386
Task: 5963