Update patch set 4
Patch Set 4:

Yes this does lead to data loss but the API shouldn't be receiving any metrics further than 2 min in the future anyway because that would point to an ntp problem. 
My previous patch did not take caching into consideration, which was a mistake. It should give the forwarder a reasonable amount of time to catch up, but I do think that the 2 weeks the Java API allows is sufficient.

Patch-set: 4