Update patch set 4
Patch Set 4:

The monasca API needs to be able to ingest data in the past as well as in the future.  The most critical reason is to avoid data loss.  E.G. an old agent that comes back online needs to be able to send the data to the API.. An external application wants to send data for threshold management, etc.  Moreover an external system (like a ML prediction) needs to be able to post metrics in the future.

All in all this feature if it gets implemented needs to be disabled by default and with both a higher and lower bound configurable.  It is unlikely that 2 minutes is a the ideal value and needs to be configurable.

Additionally, the description and the code change diverges.  The description says that any events in the past or in the future within a 2 min range will be rejected.  The code implementation rejects events 2min in the future and 2 weeks in the past.

Finally i am unsure DST would not affect the boundaries.  Basically during DST, 2 weeks is not equal to 2 x 14 x 24h, ie TWO_WEEKS = 2 * 7 * 24 * 60 * 60 * 1000 is false.

Patch-set: 4