Factor out REST utils from monasca-api
The purpose of this change is to make helper methods from the
monasca-api available in the monasca-log-api to facilitate
extending the monasca-log-api to support log queries. Only the
low hanging fruit is addressed in this commit. There are additional
methods which could also be moved here, but these would require more
substantial refactoring. These changes can be made in a follow up
commit.

Minimal changes to the refactored code have been made to make it
easier to review. Some missing unit tests will be added in follow up
commits.

Change-Id: I9831866143df39f083f9c12bc0afb10311573d29
Story: 2000934
Task: 14371