Update patch set 6
Patch Set 6: Code-Review-2

As discussed, the plan is now to keep the log-query-api change contained within the monasca-log-api repo, and perform refactoring afterwards. A spec will be written to consider merging the Monasca APIs into a single repo as an alternative to factoring out common code from the APIs to monasca-common.

Patch-set: 6
Label: Code-Review=-2