Update patch set 3
Patch Set 3: Code-Review-2

As discussed, the plan is now to keep the log-query-api change contained within the monasca-log-api repo, and perform refactoring afterwards. A spec will be written to consider merging the Monasca APIs into a single repo as an alternative to factoring out common code from the APIs to monasca-common.

Patch-set: 3
Label: Code-Review=-2