Add the metric name to the Exception messages when validation fails
When trying to debug a problem where a null value was being set for
a metric, it was very difficult to find the source because the name
of the metric was not included in the Exception message.
This commit adds the name to the validation messages for validation
methods that occur after the name has been validated.

Change-Id: I130b13d5b8a5c359063dfd67a7b934b723c7b8bf