Update patch set 2
Patch Set 2:

Not sure what the two failures are.  Here is one error message:

AssertionError: Failed test_list_metrics_with_tenant: timeout on waiting for metrics: at least one metric is needed. Current number of metrics = 0

As it is a timeout error, it may be that a recheck would get past it.

Patch-set: 2