Update patch set 1
Patch Set 1:

(1 comment)

Please undo all formatting changes and submit only the code that needs to be changed to implement the task.

Changing the formatting makes it more difficult to see the actual logic that has changed and messes up the git history. For example, it makes git blame less effective.

Patch-set: 1