Update patch set 1
Patch Set 1:

We were reluctant to changing the `kafka_lib` package in the past as it is the copy of the old version of `kafka-python` library. But with the need of having Python 3.7 support which is one of the Train release goals we don't have a choice.
Thanks for the change.

Patch-set: 1
Label: Verified=0